+++ packages/atomic/src/components/common/atomic-icon/atomic-icon.stories.tsx
@@ -1,6 +1,6 @@
+import {defaultStory} from '@coveo/atomic-storybook';
 import bgIcons from '@salesforce-ux/design-system/design-tokens/dist/bg-standard.common';
 import {html} from 'lit-html';
-import defaultStory from '../../../../.storybook/default-story';
 import AssetsList from '../../../../docs/assets.json';
 
 function snakeToCamel(value: string) {