+++ packages/atomic/src/components/search/result-template-components/atomic-result-text/atomic-result-text.stories.tsx
@@ -1,4 +1,4 @@
-import defaultResultComponentStory from '../../../../../.storybook/default-result-component-story';
+import {defaultResultComponentStory} from '@coveo/atomic-storybook';
 
 const {defaultModuleExport, exportedStory} = defaultResultComponentStory(
   'Atomic/ResultList/ResultText',